// src/types/record.ts
export interface Record {
  id: string;
  studentId: string;
  studentName: string;
  pages: string;
  reason: string; // السبب
  teacher: string; // المعلم
  dateTime: string; // التاريخ والوقت
  date: string; // التاريخ
  studentNumber: string; // رقم الطالب
  teacherName: string; // اسم المعلم
  totalPoints: string; // مجموع النقاط
  level: string; // المستوى
  badge: string; // الشارة
}